Okay Guys, HUGE Cleaning And Changes, Read Desc.
Okay, so Im in the middle of cleaning up the main mod class and making
separate registry files for blocks, items, and registry. Be careful when
calling blocks and items now because they are now declared in
ItemRegistry, and BlockRegistry. This will make the whole of the mod
files look a bit different and it may be tricky to get used to at first.
But in the long haul, it is a much better format to work on this code.
This is only the first chunk, ill be putting up the rest in a little
while.