Crafting Recipes: READ DESCRIPTION!!! (Everyone!)
Hey, just one thing...
I think the Flint Knife is a bad idea...
Because what if the player doesn't have flint, or can't get any because
the are underground?
Well, you can't play, because you can't make sticks without it...
So you have to mine every block by hand (which takes the same as mining
an obsidian block)? Well that sucks!
What if doing with the Flint Knife could give 4 sticks while just
regular recipe would be only 2 ? Wouldn't that be better ? Because it
makes it almost impossible...
The goal of this mod is not to make Minecraft so hardcore that it's
almost impossible, but to improve Minecraft and make it more
challenging...

Same thing for removing Wooden Tools... Why removing them ? They are
good the way they are! Because if you mine stone with your fist 1. It
takes a lot of time! 2. I’m not even sure it gives cobblestone… So how
could you make stone tools if you can’t get cobblestone?